Moved some of the early code about image manipulations into the folder called proofOfConcept. Modified testImageMaker to close figures after saving them (uses way less RAM) and to make the center of the square actually appear inside the image (as opposed to just edges). Made a 1000 images.... yeah I did that. Use it for playing with the Neural Net (when it works). Added a new class, TrainingImage, which inherits from InputImage and then figures out what index the square is at for the purposes of training the net.
Started working on main.py, not done by a long shot.
As of right now, the section labeled Fetch Data is the only one that works completely. It makes TrainingImages and then extracts the periphery information.
I'm not entirely sure the model is built correctly. The intent is for the net to receive a 2d input that is 100x100 in black and white and output a 10x10 2d array in black and white. So far I have added a Convolution2D layer, followed by a MaxPooling2D layer (which should take the size down to 10x10). The model is then compiled (this part used to give me errors that magically vanished). After getting the data we try to fit the data. The fit crashes into the first epoch, with an error message about argument shapes not being nice for theano.